Formatting and help message for raid command
