+++ package.json
@@ -7,6 +7,11 @@
     "start": "tsc src/index.ts; node src/index.js"
   },
   "dependencies": {
+    "axios": "^0.21.1",
     "discord.js": "^12.5.1"
+  },
+  "devDependencies": {
+    "@types/node": "^14.14.37",
+    "typescript": "^4.2.4"
   }
 }
